refactor: disable unused camera view functions
